+++ main-back/src/apps/main-gql/cron.ts
@@ -8,9 +8,9 @@ import {
   ParseTgSourceParticipantsCmdHandler,
 } from "modules/main/command/handlers/parse-tg-source-participants";
 import { TgSourceParticipantStatusDS } from "modules/main/command/projections/tg-participant-status";
-import { TgSourceDS } from "modules/main/command/projections/tg-source";
 import { TgSourceParticipantDS } from "modules/main/command/projections/tg-source-participant";
 import { TgSourceParticipantWithStatusDS } from "modules/main/command/projections/tg-source-participant-with-status";
+import { TgSourceDS } from "modules/main/command/projections/tg-source/ds";
 import { TgUserDS } from "modules/main/command/projections/tg-user";
 import { Logger } from "winston";
