+++ main-back/src/modules/ia/command/projections/index.ts
@@ -1,3 +1,3 @@
 import { UserDS } from "modules/ia/command/projections/user";
 
-export type CommonDS = UserDS;
+export type IAModuleDS = UserDS;
