+++ main-back/src/modules/main/command/projections/index.ts
@@ -1,0 +1,3 @@
+import { TgApplicationDS } from "modules/main/command/projections/tg-application";
+
+export type MainModuleDS = TgApplicationDS;
