+++ main-back/src/modules/main/command/projections/tg-homunculus/ds.ts
@@ -1,40 +1,11 @@
-import { UUID } from "fdd-ts/fop-utils";
-import {
-  BrandedPrimitive,
-  NotEmptyString,
-} from "functional-oriented-programming-ts/branded";
+import { NotEmptyString } from "functional-oriented-programming-ts/branded";
 import { Knex } from "knex";
 import { TgHomunculusTable } from "libs/main-db/models";
-
-export type TgHomunculusId = BrandedPrimitive<
-  UUID,
-  { readonly TgHomunculusId: unique symbol }
->;
-export const TgHomunculusId = {
-  new: () => {
-    return UUID.create() as TgHomunculusId;
-  },
-  ofString: (value: string) => {
-    return UUID.ofString(value) as TgHomunculusId;
-  },
-};
-
-export type TgHomunculusPhone = BrandedPrimitive<
-  NotEmptyString,
-  { readonly HomunculusPhone: unique symbol }
->;
-export const TgHomunculusPhone = {
-  ofString: (value: string) => {
-    return NotEmptyString.ofString(value) as TgHomunculusPhone;
-  },
-};
-
-export type TgHomunculus = TgHomunculusTable & {
-  id: TgHomunculusId;
-  phone: TgHomunculusPhone;
-  master: boolean;
-  authToken: NotEmptyString | null;
-};
+import {
+  TgHomunculus,
+  TgHomunculusId,
+  TgHomunculusPhone,
+} from "modules/main/command/projections/tg-homunculus/projection";
 
 const TgHomunculusDM = {
   fromTable: (table: TgHomunculusTable): TgHomunculus => {
@@ -46,7 +17,6 @@ const TgHomunculusDM = {
     };
   },
 };
-
 export type TgHomunculusDS = ReturnType<typeof TgHomunculusDS>;
 
 export const TgHomunculusDS = (deps: { knex: Knex }) => {
