+++ main-back/src/modules/main/command/projections/tg-participant-status/index.ts
@@ -1,0 +1,2 @@
+export * from "./projection";
+export * from "./ds";
