+++ main-back/src/modules/main/command/projections/tg-source-participant-with-status/ds.ts
@@ -1,27 +1,9 @@
 import { Knex } from "knex";
-import {
-  TgSourceParticipantStatusDS,
-  TgSourceParticipantStatusType,
-} from "modules/main/command/projections/tg-participant-status";
+import { TgSourceParticipantStatusDS } from "modules/main/command/projections/tg-participant-status";
 import { TgSourceId } from "modules/main/command/projections/tg-source";
-import {
-  TgSourceParticipant,
-  TgSourceParticipantDS,
-} from "modules/main/command/projections/tg-source-participant";
+import { TgSourceParticipantDS } from "modules/main/command/projections/tg-source-participant";
+import { TgSourceParticipantWithStatus } from "modules/main/command/projections/tg-source-participant-with-status/projection";
 import { TgUserId } from "modules/main/command/projections/tg-user";
-
-export type TgSourceParticipantWithStatusStatus =
-  | TgSourceParticipantStatusType
-  | "None";
-export type TgSourceParticipantWithStatus =
-  | {
-      participant: TgSourceParticipant;
-      status: TgSourceParticipantWithStatusStatus;
-    }
-  | {
-      participant: undefined;
-      status: undefined;
-    };
 
 export type TgSourceParticipantWithStatusDS = ReturnType<
   typeof TgSourceParticipantWithStatusDS
