+++ main-back/src/modules/main/command/projections/tg-user/projection.ts
@@ -3,7 +3,6 @@ import {
   BrandedPrimitive,
   NotEmptyString,
 } from "functional-oriented-programming-ts/branded";
-import { Knex } from "knex";
 import { TgUserTable } from "libs/main-db/models";
 import { Api } from "telegram";
 
@@ -21,7 +20,6 @@ export const TgUserId = {
     return UUID.ofString(value) as TgUserId;
   },
 };
-
 export type TgUserTgId = BrandedPrimitive<
   number,
   { readonly TgUserTgId: unique symbol }
@@ -31,7 +29,6 @@ export const TgUserTgId = {
     return value as TgUserTgId;
   },
 };
-
 export type TgUserUsername = BrandedPrimitive<
   NotEmptyString,
   { readonly TgUserTgTgUsername: unique symbol }
@@ -41,7 +38,6 @@ export const TgUserUsername = {
     return NotEmptyString.ofString(value) as TgUserUsername;
   },
 };
-
 export type TgUser = TgUserTable & {
   id: TgUserId;
   tgId: TgUserTgId;
@@ -98,39 +94,4 @@ export const TgUser = {
       tgLangCode: newApiUser.langCode || null,
     };
   },
-};
-
-export type TgUserDS = ReturnType<typeof TgUserDS>;
-
-export const TgUserDS = (knex: Knex) => {
-  return {
-    findByTgId: async (tgId: TgUserTgId): Promise<TgUser | undefined> => {
-      const res = await TgUserTable(knex).where("tgId", tgId).first();
-
-      return !res
-        ? undefined
-        : {
-            ...res,
-            id: res.id as TgUserId,
-            tgId: res.tgId as TgUserTgId,
-            tgUsername: res.tgUsername as TgUserUsername | null,
-            tgPhone: res.tgPhone as NotEmptyString | null,
-          };
-    },
-
-    update: async (projection: TgUser): Promise<void> => {
-      await TgUserTable(knex).where({ id: projection.id }).update(projection);
-    },
-
-    create: async (
-      projection: TgUser,
-      ignoreTgId: boolean = false
-    ): Promise<void> => {
-      if (!ignoreTgId) {
-        return TgUserTable(knex).insert(projection);
-      }
-
-      await TgUserTable(knex).insert(projection).onConflict("tgId").ignore();
-    },
-  };
 };
